requirements.txt
langgraph
langchain
langchain-openai
langchain-community
langchain-qdrant
qdrant-client
tavily-python
python-dotenv
pydantic
matplotlib
pandas
